+++ shell/components/Tabbed/index.vue
@@ -9,8 +9,8 @@ import { getApplicableExtensionEnhancements } from '@shell/core/plugin-helpers';
 import Tab from '@shell/components/Tabbed/Tab';
 import { ref } from 'vue';
 import { useIsInResourceDetailDrawer } from '@shell/components/Drawer/ResourceDetailDrawer/composables';
-import { useIsInResourceDetailPage } from '@shell/components/ResourceDetail/index.vue';
-import { useIsInResourceEditPage, useIsInResourceCreatePage } from '@shell/components/CruResource.vue';
+import { useIsInResourceDetailPage } from '@shell/composables/resourceDetail';
+import { useIsInResourceCreatePage, useIsInResourceEditPage } from '@shell/composables/cruResource';
 
 export default {
   name: 'Tabbed',
